+++ src/Light.java
@@ -2,9 +2,9 @@ package src;
 
 public class Light extends Entity
 {
-    private double transformX;
-    private double transformY;
-    private double transformZ;
+    private float transformX;
+    private float transformY;
+    private float transformZ;
 
     public Light()
     {
